Todo/pyproject.toml
[project]
name = "todo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.1",
    "fastapi[standard]>=0.115.6",
    "passlib>=1.7.4",
    "pydantic-settings>=2.7.1",
    "sqlalchemy>=2.0.37",
]
